backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
python-dotenv
supabase
